requirements.txt
Adafruit-Blinka==8.67.0
adafruit-circuitpython-busdevice==5.2.14
adafruit-circuitpython-connectionmanager==3.1.6
adafruit-circuitpython-motor==3.4.18
adafruit-circuitpython-mpu6050==1.3.5
adafruit-circuitpython-pca9685==3.4.20
adafruit-circuitpython-register==1.11.1
adafruit-circuitpython-requests==4.1.15
adafruit-circuitpython-typing==1.12.3
Adafruit-PlatformDetect==3.84.1
Adafruit-PureIO==1.1.11
binho-host-adapter==0.1.6
board==1.0
cowsay==6.1
lgpio==0.2.2.0
mpu6050-raspberrypi==1.2
numpy==2.3.4
pyftdi==0.57.1
pyPS4Controller==1.2.5
pyserial==3.5
pyusb==1.3.1
RPi.GPIO==0.7.1
smbus==1.1.post2
smbus2==0.5.0
sysv_ipc==1.1.0
typing_extensions==4.15.0
